backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
apscheduler>=3.10,<4.0
python-multipart>=0.0.9
pydantic>=2.9,<3.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
numpy>=1.26.4
python-dateutil>=2.8.2
bcrypt>=4.1
pyjwt>=2.8,<3.0
